Update dependencies and address security fixes tokio and prettytable-rs (#28)
* Update dependencies and address security fixes:
- tokio:
An issue was discovered in the tokio crate before 1.8.4, and 1.9.x through 1.13.x before 1.13.1, for Rust. In certain circumstances involving a closed oneshot channel, there is a data race and memory corruption.

- prettytable-rs:
In function Table::as_ref, a reference of vector is force cast to slice. There are multiple problems here:

    To guarantee the size is correct, we have to first do Vec::shrink_to_fit. The function requires a mutable reference, so we have to force cast from immutable to mutable, which is undefined behavior (UB).
    Even if (1) is sound, &Vec<T> and &[T] still might not have the same layout. Treating them equally may lead to undefinted behavior (UB).

* Update nix to 0.26.2 and prometheus to 0.13.

---------

Co-authored-by: Marcelo Araujo <marcelo.araujo@gandi.net>